requirements.txt
django
django-parler
django-phonenumber-field
phonenumbers